refactor: move ProfileEditForm to dedicated file
